Moved credits into the template
The theme credits were improperly hooked to wp_footer. They now exist
directly in the template. This may be changed for a template part in
the future as we introduce logic for displaying elements via the
Customizer in the footer.

I also re-ordered the THA hooks and fixed the indentation of the file.

Props to @nikolov-tmw for catching this. Fixes #21