Changed hook prioriteis
We were running the setup function a little later than we should have
been and the includes function a little earlier. This change will make
sure add_theme_support always works as expected and included files will
always have full access to Hybrid Core modules.

The only hybrid core functions which fire later than this are related
to translation and admin functionality, neither of which should be
handled in this main includes function.